Increase modularity of COS
