+++ semeio/jobs/correlated_observations_scaling/exceptions.py
@@ -1,2 +1,8 @@
 class EmptyDatasetException(ValueError):
     pass
+
+
+class ValidationError(ValueError):
+    def __init__(self, message, errors):
+        super(ValidationError, self).__init__(message)
+        self.errors = errors
