+++ semeio/jobs/scripts/spearman_correlation.py
@@ -1,8 +1,12 @@
 import argparse
 
+from ert_data.measured import MeasuredData
 from ert_shared.libres_facade import LibresFacade
 from res.enkf import ErtScript
-
+from semeio.jobs.correlated_observations_scaling.exceptions import EmptyDatasetException
+from semeio.jobs.scripts.correlated_observations_scaling import (
+    CorrelatedObservationsScalingJob,
+)
 from semeio.jobs.spearman_correlation_job.job import spearman_job
 
 
@@ -10,10 +14,22 @@ class SpearmanCorrelationJob(ErtScript):
     def run(self, *args):
         facade = LibresFacade(self.ert())
 
+        obs_keys = [
+            facade.get_observation_key(nr)
+            for nr, _ in enumerate(facade.get_observations())
+        ]
+        measured_data = MeasuredData(facade, obs_keys)
+
         parser = spearman_job_parser()
         args = parser.parse_args(args)
 
-        spearman_job(facade, args.threshold, args.dry_run)
+        scaling_configs = spearman_job(measured_data, args.threshold)
+
+        if not args.dry_run:
+            try:
+                CorrelatedObservationsScalingJob(self.ert()).run(scaling_configs)
+            except EmptyDatasetException:
+                pass
 
 
 def spearman_job_parser():
